Restructure /src/lib/ and update references
